requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
asyncstdlib==3.10.9
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
browser-cookie3==0.19.1
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
curl-cffi==0.5.10
duckduckgo-search==4.1.0
exceptiongroup==1.2.0
fastapi==0.105.0
Flask==3.0.0
frozenlist==1.4.1
g4f==0.1.9.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
Js2Py==0.74
loguru==0.7.2
lxml==4.9.4
lz4==4.3.2
magic-filter==1.0.12
MarkupSafe==2.1.3
mmh3==4.0.1
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.2
outcome==1.3.0.post0
Pillow==10.1.0
platformdirs==4.1.0
py-arkose-generator==0.0.0.2
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
PyExecJS==1.5.1
pyjsparser==2.7.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.27.0
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzlocal==5.2
undetected-chromedriver==3.5.4
urllib3==2.1.0
uvicorn==0.25.0
waitress==2.1.2
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
